Upgrade PPWCode.Server.Core.Hangfire
An attentive reader might notice that the wildcard version "1.*" is
actually identical as in the previous version of this package.  This
reader might further notice that the complete contents of this package
are actually the same as in the previous version of this package.
So the reader might go on to question the sanity of the creator of this
new version of the package.  Let it be noted that the author feels the
same way.

However, this new release is needed because the wildcard dependency
"1.*" is replaced by a dependency on the most recent version of the
dependency that matches the wildcard at the moment of package creation.
Because the dependency has been upgraded earlier on, a new upgrade of
this package is also needed, even though technically nothing in the code
was changed.